refactor(workflow): trim PR to the behavioural change
Revert test-only extractions, drop tests that re-cover shared helpers
or pre-existing behaviour, cut comment essays to one or two lines.
No behaviour change.
