+++ apps/backend/src/app/modules/form/admin-form/admin-form.middlewares.ts
@@ -73,82 +73,59 @@ export const getWebhookSettingsValidator = celebrate({
   }),
 })
 
-/**
- * Body shape for POST /forms/:formId/workflow/.
- *
- * FRM-2489: this validates **format and shape only**. Whether a step is
- * complete enough is decided in the service layer, which knows the form's
- * status — incomplete is permitted on a Private form and rejected on a Public
- * one, a distinction Joi cannot make from the request body alone.
- *
- * `field` and `conditional_field` must therefore be *omitted* rather than sent
- * empty when the admin has not chosen one. `Joi.string()` rejects `''`, and an
- * empty string would fail to cast to an ObjectId further down anyway.
- *
- * Exported separately from the `celebrate` wrapper so it can be unit tested.
- */
-export const createWorkflowStepBodyValidator = Joi.object({
-  workflow_type: Joi.string().valid(...Object.values(WorkflowType)),
-  emails: Joi.when('workflow_type', {
-    is: WorkflowType.Static,
-    // Requires the key, not a non-empty array: `[]` passes, and always has.
-    then: Joi.array().items(Joi.string().email()).required(),
-  }),
-  field: Joi.when('workflow_type', {
-    is: WorkflowType.Dynamic,
-    then: Joi.string().optional(),
-  }),
-  edit: Joi.array().items(Joi.string()).required(),
-  approval_field: Joi.string().optional(),
-  conditional_field: Joi.when('workflow_type', {
-    is: WorkflowType.Conditional,
-    then: Joi.string().optional(),
-    otherwise: Joi.forbidden(),
-  }),
-  step_name: Joi.string().optional(),
-})
-
+// FRM-2489: field / conditional_field may be omitted on a half-built step; completeness is checked in the service against form status.
 /**
  * Joi validator for POST /forms/:formId/workflow/ route.
  */
 export const createWorkflowStepValidator = celebrate({
-  [Segments.BODY]: createWorkflowStepBodyValidator,
+  [Segments.BODY]: Joi.object({
+    workflow_type: Joi.string().valid(...Object.values(WorkflowType)),
+    emails: Joi.when('workflow_type', {
+      is: WorkflowType.Static,
+      then: Joi.array().items(Joi.string().email()).required(),
+    }),
+    field: Joi.when('workflow_type', {
+      is: WorkflowType.Dynamic,
+      then: Joi.string().optional(),
+    }),
+    edit: Joi.array().items(Joi.string()).required(),
+    approval_field: Joi.string().optional(),
+    conditional_field: Joi.when('workflow_type', {
+      is: WorkflowType.Conditional,
+      then: Joi.string().optional(),
+      otherwise: Joi.forbidden(),
+    }),
+    step_name: Joi.string().optional(),
+  }),
   [Segments.PARAMS]: Joi.object({
     formId: Joi.string().required(),
   }),
 })
 
 /**
- * Body shape for PUT /forms/:formId/workflow/:stepNumber.
- * See `createWorkflowStepBodyValidator` for why completeness is not checked
- * here.
- */
-export const updateWorkflowStepBodyValidator = Joi.object({
-  _id: Joi.string().required(),
-  workflow_type: Joi.string().valid(...Object.values(WorkflowType)),
-  emails: Joi.when('workflow_type', {
-    is: WorkflowType.Static,
-    then: Joi.array().items(Joi.string().email()).required(),
-  }),
-  field: Joi.when('workflow_type', {
-    is: WorkflowType.Dynamic,
-    then: Joi.string().optional(),
-  }),
-  edit: Joi.array().items(Joi.string().hex().length(24)).required(),
-  approval_field: Joi.string().optional(),
-  conditional_field: Joi.when('workflow_type', {
-    is: WorkflowType.Conditional,
-    then: Joi.string().optional(),
-    otherwise: Joi.forbidden(),
-  }),
-  step_name: Joi.string().optional(),
-})
-
-/**
  * Joi validator for PUT /forms/:formId/workflow/:stepNumber route.
  */
 export const updateWorkflowStepValidator = celebrate({
-  [Segments.BODY]: updateWorkflowStepBodyValidator,
+  [Segments.BODY]: Joi.object({
+    _id: Joi.string().required(),
+    workflow_type: Joi.string().valid(...Object.values(WorkflowType)),
+    emails: Joi.when('workflow_type', {
+      is: WorkflowType.Static,
+      then: Joi.array().items(Joi.string().email()).required(),
+    }),
+    field: Joi.when('workflow_type', {
+      is: WorkflowType.Dynamic,
+      then: Joi.string().optional(),
+    }),
+    edit: Joi.array().items(Joi.string().hex().length(24)).required(),
+    approval_field: Joi.string().optional(),
+    conditional_field: Joi.when('workflow_type', {
+      is: WorkflowType.Conditional,
+      then: Joi.string().optional(),
+      otherwise: Joi.forbidden(),
+    }),
+    step_name: Joi.string().optional(),
+  }),
   [Segments.PARAMS]: Joi.object({
     formId: Joi.string().required(),
     stepNumber: Joi.number().integer().min(0).required(),
