+++ apps/frontend/src/features/admin-form/create/workflow/components/WorkflowContent/EditStepBlock/EditStepBlock.tsx
@@ -2,7 +2,11 @@ import { useEffect, useLayoutEffect, useRef } from 'react'
 import { useForm } from 'react-hook-form'
 import { Box, Stack } from '@chakra-ui/react'
 
-import { FormWorkflowStep } from 'formsg-shared/types'
+import {
+  FormWorkflowStep,
+  FormWorkflowStepBase,
+  WorkflowType,
+} from 'formsg-shared/types'
 
 import { SaveActionGroup } from '~features/admin-form/create/logic/components/LogicContent/EditLogicBlock/EditCondition'
 import { useUser } from '~features/user/queries'
@@ -24,7 +28,6 @@ import { SpotlightGroup } from '../../Spotlight'
 import { isFirstStepByStepNumber } from '../utils/isFirstStepByStepNumber'
 
 import { ApprovalsBlock } from './ApprovalsBlock'
-import { buildWorkflowStep } from './buildWorkflowStep'
 import { GuidedActionGroup } from './GuidedActionGroup'
 import { QuestionsBlock } from './QuestionsBlock'
 import { RespondentBlock } from './RespondentBlock'
@@ -45,6 +48,80 @@ export const FIELDS_TO_EDIT_NAME = 'edit'
 export const APPROVAL_FIELD_NAME = 'approval_field'
 
 const SECTION_REVEAL_SCROLL_DELAY_MS = 100
+
+/**
+ * Builds a workflow step from form inputs, or undefined if they cannot form a
+ * valid step. Inputs are assumed already validated (both save paths run
+ * handleSubmit first), so the field narrowing here is a type guarantee, not the
+ * validation gate.
+ */
+export const buildWorkflowStep = (
+  rawInputs: EditStepInputs,
+  isFirstStep: boolean,
+): (FormWorkflowStep & { _id: string }) | undefined => {
+  const inputs = { ...rawInputs }
+  if (inputs.approval_field === '') {
+    inputs.approval_field = undefined
+  }
+  if (inputs.step_name === '') {
+    inputs.step_name = undefined
+  }
+
+  if (isFirstStep) {
+    return inputs.field
+      ? {
+          ...inputs,
+          workflow_type: WorkflowType.Dynamic,
+          field: inputs.field,
+        }
+      : {
+          ...inputs,
+          workflow_type: WorkflowType.Static,
+          emails: inputs.emails ?? [],
+        }
+  }
+
+  const workflowStepBase: FormWorkflowStepBase & { _id: string } = {
+    _id: inputs._id,
+    workflow_type: inputs.workflow_type,
+    edit: inputs.edit,
+    approval_field: inputs.approval_field,
+    step_name: inputs.step_name,
+  }
+
+  switch (inputs.workflow_type) {
+    case WorkflowType.Static: {
+      return {
+        ...workflowStepBase,
+        // Need to explicitly set workflow_type in this object to help with typechecking.
+        workflow_type: WorkflowType.Static,
+        emails: inputs.emails ?? [],
+      }
+    }
+    // FRM-2489: an unchosen respondent is omitted, not sent as ''.
+    case WorkflowType.Dynamic: {
+      return {
+        ...workflowStepBase,
+        workflow_type: WorkflowType.Dynamic,
+        ...(inputs.field ? { field: inputs.field } : {}),
+      } as FormWorkflowStep & { _id: string }
+    }
+    case WorkflowType.Conditional: {
+      return {
+        ...workflowStepBase,
+        workflow_type: WorkflowType.Conditional,
+        ...(inputs.conditional_field
+          ? { conditional_field: inputs.conditional_field }
+          : {}),
+      } as FormWorkflowStep & { _id: string }
+    }
+    default: {
+      // Exhaustiveness check: a new WorkflowType breaks the build here until handled.
+      const exhaustiveCheck: never = inputs
+      return exhaustiveCheck
+    }
+  }
+}
 
 export const EditStepBlock = ({
   stepNumber,
