+++ apps/frontend/src/features/admin-form/create/workflow/components/WorkflowContent/EditStepBlock/RespondentBlock/Components/DynamicRespondentOption.tsx
@@ -70,8 +70,6 @@ export const DynamicRespondentOption = ({
               control={control}
               name="field"
               rules={{
-                // FRM-2489: a respondent is still needed before the form goes
-                // live, just not before it can be saved.
                 required: isSavePermissive
                   ? false
                   : t(
