+++ apps/frontend/src/features/admin-form/create/workflow/hooks/useIsWorkflowSavePermissive.ts
@@ -5,13 +5,8 @@ import { useAdminForm } from '~features/admin-form/common/queries'
 import { useIsWorkflowBuilderRedesign } from './useIsWorkflowBuilderRedesign'
 
 /**
- * Whether a half-built workflow step may be saved as-is (FRM-2489).
- *
- * Two conditions. Whether the *form* may hold an incomplete workflow is the
- * server's call, deferred to here through the shared policy. The redesign flag
- * is an additional client-side condition: the old builder keeps its existing
- * inline validation, so rolling the flag back restores the old editing
- * experience even though the server would still accept the save.
+ * Whether a half-built workflow step may be saved as-is (FRM-2489). Defers to
+ * the shared server policy, gated additionally on the redesign flag being on.
  */
 export const useIsWorkflowSavePermissive = (): boolean => {
   const { data: form } = useAdminForm()
