+++ packages/shared/utils/workflow-step-completion.ts
@@ -4,24 +4,8 @@ import { FormStatus, FormWorkflowStep, WorkflowType } from '../types/form'
 import { checkIsOptionsMismatched } from './options-recipients-map-validation'
 
 /**
- * Whether a workflow has to be complete before it may be saved.
- *
- * Only a live form needs a runnable workflow. Building one is not linear: an
- * admin often knows which fields a step needs before they know who fills them
- * in, and a form nobody can submit to is safe to leave half-built.
- *
- * Deliberately does **not** consult the redesign flag. An earlier version made
- * flag-off strict for every status, reasoning that a rollback should not leave
- * a live form unguarded. But this ticket is what introduced the check at all:
- * before it, no layer required a Static step to have a recipient. Applying the
- * new rule to Private forms under flag-off would therefore reject workflow
- * mutations on forms that saved legally, and since the guard checks the whole
- * resulting workflow, a form with two such steps could not be repaired at all.
- * Gating on `Public` keeps live forms guarded in both flag states, which was
- * the actual concern.
- *
- * An unknown status is treated as not-live. The server always knows the status;
- * on the client this only defers to the server's answer on save.
+ * Only a live form needs a runnable workflow. Deliberately not gated on the
+ * redesign flag; see PR #9856.
  */
 export const mustWorkflowBeComplete = ({
   formStatus,
@@ -29,11 +13,7 @@ export const mustWorkflowBeComplete = ({
   formStatus?: FormStatus
 }): boolean => formStatus === FormStatus.Public
 
-/**
- * Conditional routing is only usable once every option of the selected dropdown
- * has at least one recipient, otherwise a real submission routes nowhere. The
- * mapping lives on the dropdown field, so this needs the form's fields.
- */
+// Every option of the selected dropdown needs at least one recipient, or a real submission routes nowhere.
 const isConditionalRoutingComplete = (
   // Optional at runtime on a half-built step, though the type says otherwise.
   conditionalFieldId: FormFieldDto['_id'] | undefined,
@@ -65,15 +45,8 @@ const isConditionalRoutingComplete = (
 }
 
 /**
- * Whether a workflow step is complete enough to run.
- *
- * A step needs a person to fill it in, but not fields to fill: a respondent who
- * only reads the submission and passes it on is legitimate, and live forms do
- * this. Step 0 is exempt from needing a person, since its respondent is always
- * whoever opens the form.
- *
- * Ids are compared as strings: the frontend holds them as strings and Mongoose
- * hands them over as ObjectIds, where `===` compares identity and never matches.
+ * Whether a workflow step is complete enough to run. Needs a person, not fields;
+ * step 0 is exempt. Ids are compared as strings since Mongoose hands back ObjectIds.
  *
  * @param stepNumber the step's position in the full, unfiltered workflow
  */
@@ -82,8 +55,7 @@ export const isStepComplete = (
   formFields: FormFieldDto[],
   stepNumber: number,
 ): boolean => {
-  // An approval field the assigned person cannot see leaves them nothing to
-  // answer, and the workflow stops at their step permanently.
+  // An approval field the assigned person cannot see stalls the workflow permanently.
   if (
     step.approval_field &&
     !step.edit.map(String).includes(String(step.approval_field))
@@ -103,10 +75,7 @@ export const isStepComplete = (
   }
 }
 
-/**
- * The indices of every incomplete step in a workflow, in order. Callers use
- * these to name the offending steps to the admin.
- */
+// The indices of every incomplete step in a workflow, in order; used to name the offending steps to the admin.
 export const getIncompleteStepNumbers = (
   workflow: FormWorkflowStep[],
   formFields: FormFieldDto[],
